Refactoring some of the components
