+++ client-app/src/components/BookmarkFilter.ts
@@ -1,0 +1,7 @@
+import Vue from "vue";
+import Component from "vue-class-component";
+
+@Component
+export default class BookmarkFilter extends Vue {
+
+}
