+++ client-app/src/components/BookmarkFilter.vue
@@ -1,0 +1,14 @@
+<template>
+  <section>
+    <v-card elevation="8">
+      <v-card-title primary-title>
+        <h1>Filters</h1>
+      </v-card-title>
+      <v-card-text>
+        <p>Under construction...</p>
+      </v-card-text>
+      <v-overlay :value="true" absolute></v-overlay>
+    </v-card>
+  </section>
+</template>
+<script lang="ts" src="./BookmarkFilter.ts"></script>
